Rename files. Duplicate logic from active calculateresponse to the WIP one.
