+++ Toolbox/arcsdm/calculateresponse.py
@@ -23,78 +23,47 @@
 #  Modifications for use of Lookup_sa by Ryan Bryn, ESRI
 # ---------------------------------------------------------------------------
 """
+
+import arcgisscripting
 import arcpy
-import importlib
+import gc
 import math
 import os
 import sys
 import traceback
 
 from arcsdm.common import log_arcsdm_details
-from arcsdm.wofe_common import check_wofe_inputs, get_study_area_parameters, WofeInputError
+from arcsdm.missingdatavar_func import create_missing_data_variance_raster
+from arcsdm.wofe_common import apply_mask_to_raster, check_wofe_inputs, extract_layer_from_raster_band, get_study_area_parameters, WofeInputError
 
 
-# NOTE: WIP - trying to replace all references to gp with arcpy
-# Replaces the function MissingDataVariance from missingdatavar_func.py
-def create_missing_data_variance_layer(nodata_value, study_area_size_sq_km, weights_rasters_list, post_probability_raster, md_variance_output_name):
-    # TODO: Note! Input rasters should be masked already? - possibly add check?
-    # TODO: decide whether the weights rasters need to already be masked in a way that missing data has been combined to nodata?
-    
-    for raster in weights_rasters_list:
-        arcpy.AddMessage(f"Missing data Variance for: {raster}")
-        weights_raster = arcpy.Describe(raster).catalogPath
-
-        # Start MD Variance raster
-        # Get posterior probability at MD cells
-        R1 = os.path.join(arcpy.env.scratchWorkspace, "R1")
-        if arcpy.Exists(R1):
-            arcpy.management.Delete(R1)
-
-        temp_nodata_mask = arcpy.sa.IsNull(weights_raster)
-        pprb_at_nodata_cells = arcpy.sa.Con(temp_nodata_mask, post_probability_raster, 0.0, "VALUE = 1")
-
-        pprb_at_nodata_cells.save(R1)
-        
-        # Get PostODDs at MD cells
-        R2 = os.path.join(arcpy.env.scratchWorkspace, "R2")
-        if arcpy.Exists(R2):
-            arcpy.management.Delete(R2)
-
-        # Exp = "%s / (1.0 - %s)" % (R1, R1)
-
-        variable_names = ['"r1"']
-        post_odds_expression = "%s / (1.0 - %s)" % ('"r1"', '"r1"')
-        arcpy.AddMessage(f"R2 = {post_odds_expression}")
-        post_odds = arcpy.sa.RasterCalculator([R1], variable_names, post_odds_expression, extent_type="UnionOf", cellsize_type="MinOf")
-        post_odds.save(R2)
-        arcpy.AddMessage(f"R2 exists: {arcpy.Exists(R2)}")
-
-        # Get Total Variance of MD cells
-        # Create total class variances list
-        ClsVars = []
-
-        # TODO: for each class, calculate the variance of the posterior probability due to missing data
-
-        # TODO: 1: calculate for present pattern as if missing pattern is known:
-        # (p(D|Ej)-p(D))^2 * p(Ej)
-        # TODO: 2: calculate for absent pattern as if missing pattern is known:
-        # (p(D|nEj)-p(D))^2 * p(nEj)
-        # (ie. use the full mask area as the total area!)
-
-        # Need:
-        # p(Ej): area of predictor pattern j / total area
-        # p(nEj): area of missing data for pattern j / total area
-        # p(D): posterior probability calculated for the redion where Ej is missing -> pprb_at_nodata_cells
-        # p(D|Ej): posterior probability in the presence of pattern Ej
-        # p(D|nEj): posterior probability in the absence of pattern Ej
-        # 
-
-    return None
+debuglevel = 0
 
 
-# NOTE: WIP - trying to replace all references to gp with arcpy
+def testdebugfile():
+    returnvalue = 0 # This because python sucks in detecting outputs from functions
+    import sys
+    import os
+    if (debuglevel > 0):
+        return 1
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+    if os.path.isfile(dir_path + "/DEBUG"):
+        return 1
+    return returnvalue
+
+
+def dwrite(message):
+    debug = testdebugfile()
+    if (debuglevel > 0 or debug > 0):
+        arcpy.AddMessage("Debug: " + message)
+
+
 def Execute(self, parameters, messages):
     try:
+        # TODO: Refactor to arcpy.
+        # Some references to arcgisscripting still remain due to difficulty updating raster calculations.
+        gp = arcgisscripting.create()
+
         arcpy.CheckOutExtension("Spatial")
 
         arcpy.env.overwriteOutput = True
@@ -114,18 +83,17 @@ def Execute(self, parameters, messages):
         output_mdvar_raster = parameters[8].valueAsText
         output_total_std_raster = parameters[9].valueAsText
         output_confidence_raster = parameters[10].valueAsText
-
-        nodata_value = missing_data_value
-        if is_ignore_missing_data_selected:
-            nodata_value = "#"
+        
+        if is_ignore_missing_data_selected: # for nodata argument to CopyRaster tool
+            nodata_value = missing_data_value
+        else:
+            nodata_value = '#'
 
         input_rasters = input_evidence_rasters.split(";")
         weights_tables = input_weights_tables.split(";")
 
         if len(input_rasters) != len(weights_tables):
-            raise WofeInputError("The number of evidence rasters should equal the number of weights tables!")
-
-        # TODO: Add check for weights table columns depending on weights type (unique weights shouldn't have generalized columns)
+            raise ValueError("The number of evidence rasters should equal the number of weights tables!")
 
         check_wofe_inputs(input_rasters, training_points_feature)
 
@@ -135,37 +103,52 @@ def Execute(self, parameters, messages):
             if ("weight" not in fields) or ("w_std" not in fields):
                 raise WofeInputError(f"The weights table {weights_table} has not been generalized! Make sure each table has the columns 'WEIGHT' and 'W_STD'.")
 
-        # TODO: check that all evidence rasters have the same cell size?
-        # TODO: use the env Cell Size instead. not all of the evidence rasters necessarily have the same cell size
-        # TODO: evidence rasters should be resampled to env Cell Size?
-        evidence_cellsize = arcpy.Describe(input_rasters[0]).MeanCellWidth
-
         log_arcsdm_details()
         total_area_sq_km_from_mask, training_point_count = get_study_area_parameters(unit_cell_area_sq_km, training_points_feature)
+        
         area_cell_count = total_area_sq_km_from_mask / unit_cell_area_sq_km
+        arcpy.AddMessage(("%-20s %s" % ("Study Area:", str(area_cell_count))))
+        arcpy.AddMessage("%-20s %s" % ("# training points:", str(training_point_count)))
+
         prior_probability = training_point_count / area_cell_count
-
-        arcpy.AddMessage("\n" + "=" * 21 + " Starting calculate response " + "=" * 21)
         arcpy.AddMessage("%-20s %s"% ("Prior probability:" , str(prior_probability)))
-        arcpy.AddMessage(f"Input evidence rasters: {input_rasters}")
 
-        workspace_type = arcpy.Describe(arcpy.env.workspace).workspaceType
-
-        # TODO: check that the order of the evidence rasters and the associated weights tables matches, e.g. by checking the number of classes
-
+        arcpy.AddMessage(f"Input rasters: {input_rasters}")
+        
+        # Create weight raster from raster's associated weights table
         tmp_weights_rasters = []
         tmp_std_rasters = []
-        tmp_rasters_with_missing_data = []
+        # Create a list for the Missing Data Variance tool.
+        rasters_with_missing_data = []
+        # Evidence rasters should have missing data values, where necessary, for
+        # NoData cell values within study area.
+        # For each input_raster create a weights raster from the raster and its weights table.
+
+        workspace_type = arcpy.Describe(arcpy.env.workspace).workspaceType
+        
+        arcpy.AddMessage("\nCreating tmp weight and STD rasters...")
+        arcpy.AddMessage("=" * 41)
 
         i = 0
-
-        # Create Weight and Standard Deviation rasters
         while i < len(input_rasters):
             input_raster = input_rasters[i]
             weights_table = weights_tables[i]
 
-            arcpy.AddMessage(f"Processing evidence layer {input_raster} and weights table {weights_table}")
+            # Check each Input Raster datatype and coordinate system
+            inputDescr = arcpy.Describe(input_raster)
+            inputCoord = inputDescr.spatialReference.name
+            arcpy.AddMessage(f"{input_raster}, Data type: {inputDescr.dataType}, Coordinate System: {inputCoord}")
+            
+            input_raster, inputDescr = extract_layer_from_raster_band(input_raster, inputDescr)
 
+            trainingDescr = arcpy.Describe(training_points_feature)
+            trainingCoord = trainingDescr.spatialReference.name
+
+            if inputCoord != trainingCoord:
+                raise WofeInputError(f"ERROR: Coordinate System of Input Raster is {inputCoord} and Training points it is {trainingCoord}. These must be same.")
+
+            # When workspace type is File System, Input Weight Table also must end with .dbf
+            # If using GDB database, remove numbers and underscore from the beginning of the name (else block)
             if workspace_type == "FileSystem":
                 if not weights_table.endswith(".dbf"):
                     weights_table += ".dbf"
@@ -174,112 +157,220 @@ def Execute(self, parameters, messages):
                 while len(wtsbase) > 0 and (wtsbase[:1] <= "9" or wtsbase[:1] == "_"):
                     wtsbase = wtsbase[1:]
                 weights_table = os.path.dirname(weights_table) + "\\" + wtsbase
+
+            # Compare workspaces to make sure they match
             
-            # TODO: check if shortening the name is necessary
-            # TODO: make sure these result in unique names if evidence rasters have similar names
+            arcpy.AddMessage(f"Processing {input_raster}")
+            
             w_raster_name = input_raster.replace(".", "_")
-            w_raster_name = w_raster_name + "_W"
+            w_raster_name = w_raster_name[:10] + "W"
+            # Needs to be able to extract input raster name from full path.
+            # Can't assume only a layer from ArcMap.
+            std_raster_name = os.path.basename(input_raster[:9]).replace(".", "_") + "S" # No . allowed in filegeodatgabases
 
-            std_raster_name = input_raster.replace(".", "_")
-            std_raster_name = std_raster_name + "_S"
-
+            # If using GDB database, remove numbers and underscore from the beginning of the names
             if workspace_type != "FileSystem":
                 while len(w_raster_name) > 0 and (w_raster_name[:1] <= "9" or w_raster_name[:1] == "_"):
                     w_raster_name = w_raster_name[1:]
                 while len(std_raster_name) > 0 and (std_raster_name[:1] <= "9" or std_raster_name[:1] == "_"):
                     std_raster_name = std_raster_name[1:]
 
-            # Check for unsigned integer raster - cannot have negative missing data
-            if (nodata_value != "#") and (arcpy.Describe(input_raster).pixelType.upper().startswith("U")):
-                nodata_value = "#"
+            # Create the _W & _S rasters that will be used to calculate output rasters
+            output_tmp_w_raster = arcpy.CreateScratchName(w_raster_name, "", "rst", arcpy.env.scratchWorkspace)
+            output_tmp_std_raster = arcpy.CreateScratchName(std_raster_name, "", "rst", arcpy.env.scratchWorkspace)
             
-            # Get input raster as an in-memory raster layer, since AddJoin requires a layer
-            tmp_raster_layer = "tmp_rst"
-            arcpy.management.MakeRasterLayer(input_raster, tmp_raster_layer)
+            # Increase the count for next round
+            i += 1
+     
+            # Need to create in-memory Raster Layer for Join
+            # Check for unsigned integer raster; cannot have negative missing data
+            if nodata_value != "#" and arcpy.Describe(input_raster).pixelType.upper().startswith("U"):
+                NoDataArg2 = "#"
+            else:
+                NoDataArg2 = nodata_value
+
+            masked_evidence_raster = apply_mask_to_raster(inputDescr.catalogPath, missing_data_value)
+
+            # Create new rasterlayer from input raster for both the weights and the std raster -> Result RasterLayer
+            # These will be in-memory only
+            # AddJoin requires an input layer or tableview not Input Raster Dataset.
+            tmp_raster_layer = "OutRas_lyr"
+
+            arcpy.management.MakeRasterLayer(masked_evidence_raster, tmp_raster_layer)
+            
+            # Join result layer with weights table
             arcpy.management.AddJoin(tmp_raster_layer, "VALUE", weights_table, "CLASS")
 
-            output_tmp_w_raster = arcpy.CreateScratchName(w_raster_name, "", "RasterDataset", arcpy.env.scratchWorkspace)
-            output_tmp_std_raster = arcpy.CreateScratchName(std_raster_name, "", "RasterDataset", arcpy.env.scratchWorkspace)
+            temp_raster = arcpy.CreateScratchName("tmp_rst", "", "rst", arcpy.env.scratchWorkspace)
+            
+            # Delete existing temp_raster
+            if arcpy.Exists(temp_raster):
+                arcpy.management.Delete(temp_raster)
+                gc.collect()
+                arcpy.management.ClearWorkspaceCache()
+                arcpy.AddMessage("Deleted tempraster")
+            
+            # Copy created and joined in-memory raster to temp_raster
+            arcpy.CopyRaster_management(tmp_raster_layer, temp_raster, "#", "#", NoDataArg2)
 
-            arcpy.management.CopyRaster(tmp_raster_layer, output_tmp_w_raster, "#", "#", nodata_value)
-            arcpy.management.CopyRaster(tmp_raster_layer, output_tmp_std_raster, "#", "#", nodata_value)
-
-            weight_lookup = arcpy.sa.Lookup(output_tmp_w_raster, "WEIGHT")
-            std_lookup = arcpy.sa.Lookup(output_tmp_std_raster, "W_STD")
-
-            # Note: the step in the old code where the mask gets used is the Lookup function
-            # TODO: go properly through the old code to see if mask should be applied earlier
-            # (it might affect the logic itself, and possibly the performance)
+            arcpy.AddMessage(f"Output tmp weights raster: {output_tmp_w_raster}")
+            
+            # Save weights raster
+            weight_lookup = arcpy.sa.Lookup(temp_raster, "WEIGHT")
             weight_lookup.save(output_tmp_w_raster)
-            std_lookup.save(output_tmp_std_raster)
-
+            
+            if not arcpy.Exists(output_tmp_w_raster):
+                raise WofeInputError(f"{output_tmp_w_raster} does not exist.")
+            
             tmp_weights_rasters.append(output_tmp_w_raster)
+
+            std_lookup = arcpy.sa.Lookup(temp_raster, "W_STD")
+            std_lookup.save(output_tmp_std_raster)
+            
+            if not arcpy.Exists(output_tmp_std_raster):
+                raise WofeInputError(f"{output_tmp_std_raster} does not exist.")
+            
             tmp_std_rasters.append(output_tmp_std_raster)
 
+            # Check for Missing Data in raster's Wts table
             if not is_ignore_missing_data_selected:
-                clause = "Class = %s" % missing_data_value
-                with arcpy.da.SearchCursor(weights_table, ["Class"], where_clause=clause) as cursor:
-                    for row in cursor:
-                        if row:
-                            tmp_rasters_with_missing_data.append(output_tmp_w_raster)
-                            break
+                # Update the list for Missing Data Variance Calculation
+                tblrows = gp.SearchCursor(weights_table, "Class = %s" % missing_data_value)
+                tblrow = tblrows.Next()
+                if tblrow:
+                    rasters_with_missing_data.append(arcpy.Describe(output_tmp_w_raster).catalogPath)
+            
+            arcpy.management.Delete(temp_raster)
 
-            i += 1
-
-        arcpy.AddMessage("Finished creating tmp rasters")
-
-        arcpy.AddMessage("\Creating Post Probability Raster...\n" + "=" * 41)
-
-        variable_names = [f'"e{i}"' for i in range(len(tmp_weights_rasters))]
-        weights_sum_expression = " + ".join(i for i in variable_names)
-        prior_logit = math.log(prior_probability / (1.0 - prior_probability))
-
-        if len(tmp_weights_rasters) == 1:
-            posterior_logit_expression = "%s + %s" % (prior_logit, weights_sum_expression)
-        else:
-            posterior_logit_expression = "%s + (%s)" % (prior_logit, weights_sum_expression)
+            arcpy.AddMessage(" ") # Cycle done - add ONE linefeed
         
-        arcpy.AddMessage(f"Variables: {variable_names}")
-        arcpy.AddMessage("Posterior logit expression: " + posterior_logit_expression)
+        # Get Post Logit Raster
+        arcpy.AddMessage("\nGetting Post Logit raster...\n" + "=" * 41)
 
-        posterior_probability_expression = "Float(Exp(%s) / (1.0 + Exp(%s)))" % (posterior_logit_expression, posterior_logit_expression)
-        arcpy.AddMessage(f"Posterior probability expression: {posterior_probability_expression}")
+        Input_Data_Str = ' + '.join('"{0}"'.format(w) for w in tmp_weights_rasters)
+        arcpy.AddMessage(f"Input_data_str: {Input_Data_Str}")
 
-        # NOTE: Conflicting info about the use of RasterCalculator in Esri's documentation
-        # According to a how-to article on raster calculation, RasterCalculator isn't intended for use in scripting environments.
-        # (See: https://support.esri.com/en-us/knowledge-base/how-to-perform-raster-calculation-using-arcpy-000022418)
-        # But the arcpy RasterCalculator page does not mention anything about this.
-        # (See: https://pro.arcgis.com/en/pro-app/latest/arcpy/spatial-analyst/raster-calculator.htm)
-        posterior_probability_result = arcpy.sa.RasterCalculator(tmp_weights_rasters, variable_names, posterior_probability_expression, extent_type="UnionOf", cellsize_type="MinOf")
-        posterior_probability_result.save(output_pprb_raster)
+        Constant = math.log(prior_probability / (1.0 - prior_probability))
+        
+        if len(tmp_weights_rasters) == 1:
+            InExpressionPLOG = "%s + %s" % (Constant, Input_Data_Str)
+        else:
+            InExpressionPLOG = "%s + (%s)" % (Constant, Input_Data_Str)
 
-        # Due to bit depth issue, the resulting raster has pixel depth 64 Bit
-        # See: https://community.esri.com/t5/arcgis-spatial-analyst-questions/controling-bit-depth-of-raster-in-fgdb-format/td-p/125850
+        arcpy.AddMessage(f"InexpressionPlog: {InExpressionPLOG}")
+
+        # Get Post Probability Raster
+        arcpy.AddMessage("\nCreating Post Probability Raster...\n" + "=" * 41)
+
+        try:
+            InExpression = "Exp(%s) / (1.0 + Exp(%s))" % (InExpressionPLOG, InExpressionPLOG)
+            arcpy.AddMessage(f"InExpression = {InExpression}")
+            arcpy.AddMessage(f"Postprob: {output_pprb_raster}")
+
+            gp.SingleOutputMapAlgebra_sa(InExpression, output_pprb_raster)
+            
+            arcpy.SetParameterAsText(6, output_pprb_raster)
+        except:
+            arcpy.AddError(arcpy.GetMessages(2))
+            raise
+        else:
+            arcpy.AddWarning(arcpy.GetMessages(1))
+            arcpy.AddMessage(arcpy.GetMessages(0))
 
         arcpy.AddMessage("\nCreating Post Probability STD Raster...\n" + "=" * 41)
         
+        #TODO: Figure out what this does!? TR
+        #TODO: This is always false now
         if len(tmp_std_rasters) == 1:
-            # TODO: Just save the Std raster as is
-            pass
+            InExpression = '"%s"' % (tmp_std_rasters[0])
         else:
-            std_input_rasters = [arcpy.Describe(output_pprb_raster).catalogPath] + [arcpy.Describe(s).catalogPath for s in tmp_std_rasters]
-
-            std_variable_names = [f'"e{i}"' for i in range(len(tmp_std_rasters))]
-            variable_names = ['"pprob"'] + std_variable_names
-            std_sum_expression = " + ".join(f'SQR({i})' for i in std_variable_names)
-            constant = 1.0 / training_point_count
-            std_expression = 'SQRT(SQR("pprob") * (%s + SUM(%s)))' % (constant, std_sum_expression)
+            SUM_args_list = []
+            for Std_Raster in tmp_std_rasters:
+                SUM_args_list.append("SQR(\"%s\")" % Std_Raster)
             
-            arcpy.AddMessage(f"Posterior probability STD expression: {std_expression}")
-            arcpy.AddMessage(f"Input rasters: {std_input_rasters}")
-            arcpy.AddMessage(f"Variable names: {variable_names}")
+            SUM_args = " + ".join(SUM_args_list)
 
-            # TODO: Continue
-            # Creating the Std raster isn't working at the moment.
-            # Possibly the issue is with the pixel depth and/or cell alignment of the posterior probability raster
-            # Complains about missing VAT, but that is unlikely to be the real issue.
-            std_result = arcpy.sa.RasterCalculator(std_input_rasters, variable_names, std_expression, extent_type="UnionOf", cellsize_type="MinOf")
-            std_result.save(output_std_raster)
+            arcpy.AddMessage("Sum_args: " + SUM_args + "\n" + "=" * 41)
+       
+            Constant = 1.0 / float(training_point_count)
+
+            InExpression = "SQRT(SQR(%s) * (%s + SUM(%s)))" % (output_pprb_raster, Constant, SUM_args)
+            arcpy.AddMessage(f"InExpression = {InExpression}")
+
+        try:
+            arcpy.AddMessage(f"InExpression 2 ====> {InExpression}")
+
+            gp.SingleOutputMapAlgebra_sa(InExpression, output_std_raster)
+
+            arcpy.SetParameterAsText(7, output_std_raster)
+        except:
+            arcpy.AddError(arcpy.GetMessages(2))
+            raise
+        else:
+            arcpy.AddWarning(arcpy.GetMessages(1))
+            arcpy.AddMessage(arcpy.GetMessages(0))
+        
+        # Create Variance of missing data here and create totVar = VarMD + SQR(VarWts)
+        if not is_ignore_missing_data_selected:
+            # Calculate Missing Data Variance
+            if len(rasters_with_missing_data) > 0:
+                arcpy.AddMessage("Calculating Missing Data Variance...")
+                try:
+                    if arcpy.Exists(output_mdvar_raster):
+                        arcpy.management.Delete(output_mdvar_raster)
+
+                    create_missing_data_variance_raster(gp, rasters_with_missing_data, output_pprb_raster, output_mdvar_raster)
+
+                    InExpression = 'SQRT(SUM(SQR(%s),%s))' % (output_std_raster, output_mdvar_raster)
+
+                    arcpy.AddMessage("Calculating Total STD...")
+                    arcpy.AddMessage(f"InExpression 3 ====> {InExpression}")
+
+                    gp.SingleOutputMapAlgebra_sa(InExpression, output_total_std_raster)
+
+                    arcpy.SetParameterAsText(9, output_total_std_raster)
+                except:
+                    arcpy.AddError(arcpy.GetMessages(2))
+                    raise
+            else:
+                arcpy.AddWarning("No evidence with missing data. Missing Data Variance not calculated.")
+                output_mdvar_raster = None
+                output_total_std_raster = output_std_raster
+        else:
+            arcpy.AddWarning("Missing Data Ignored. Missing Data Variance not calculated.")
+            output_mdvar_raster = None
+            output_total_std_raster = output_std_raster
+        
+        # Confidence is PP / sqrt(totVar)
+        arcpy.AddMessage("\nCalculating Confidence...\n" + "=" * 41)
+
+        InExpression = "%s / %s" % (output_pprb_raster, output_std_raster)
+        arcpy.AddMessage(f"InExpression 4====> {InExpression}")
+        try: 
+            gp.SingleOutputMapAlgebra_sa(InExpression, output_confidence_raster)
+            arcpy.SetParameterAsText(10, output_confidence_raster)
+        except:
+            arcpy.AddError(arcpy.GetMessages(2))
+            raise
+        else:
+            arcpy.AddWarning(arcpy.GetMessages(1))
+            arcpy.AddMessage(arcpy.GetMessages(0))
+        
+        # Set derived output parameters
+        arcpy.SetParameterAsText(6, output_pprb_raster)
+        arcpy.SetParameterAsText(7, output_std_raster)
+
+        if output_mdvar_raster and (not is_ignore_missing_data_selected):
+            arcpy.SetParameterAsText(8, output_mdvar_raster)
+        else:
+            arcpy.AddWarning("No Missing Data Variance.")
+
+        if not (output_total_std_raster == output_std_raster):
+            arcpy.SetParameterAsText(9, output_total_std_raster)
+        else:
+            arcpy.AddWarning("Total STD same as Post Probability STD.")
+
+        arcpy.SetParameterAsText(10, output_confidence_raster)
 
         arcpy.AddMessage("Deleting tmp rasters...")
         for raster in tmp_weights_rasters:
@@ -287,16 +378,27 @@ def Execute(self, parameters, messages):
         
         for raster in tmp_std_rasters:
             arcpy.management.Delete(raster)
-        
-    except arcpy.ExecuteError:
+
+        arcpy.AddMessage("Done\n" + "=" * 41)
+    except arcpy.ExecuteError as e:
+        arcpy.AddError("\n")
+        arcpy.AddMessage("Calculate Response caught arcpy.ExecuteError ")
         arcpy.AddError(arcpy.GetMessages(2))
-    except Exception:
+        if len(e.args) > 0:
+            args = e.args[0]
+            args.split("\n")
+                    
+        arcpy.AddMessage("-------------- END EXECUTION ---------------")
+        raise 
+    except:
         tb = sys.exc_info()[2]
         tbinfo = traceback.format_tb(tb)[0]
 
         pymsg = "PYTHON ERRORS:\nTraceback Info:\n" + tbinfo + "\nError Info:\n    " + \
             str(sys.exc_info()) + "\n"
         msgs = "GP ERRORS:\n" + arcpy.GetMessages(2) + "\n"
-
+        
         arcpy.AddError(msgs)
         arcpy.AddError(pymsg)
+
+        raise
